requirements.txt
python-dotenv~=1.0.1
langchain~=0.3.3
langchain-community~=0.3.1
unstructured[md]
llama-parse
python-magic-bin==0.4.14
langchain-experimental